requirements.txt
praw>=7.8.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pandas>=1.5.0
matplotlib>=3.5.0
sentence-transformers>=2.2.0
scikit-learn>=1.1.0
bertopic>=0.15.0
numpy>=1.21.0
zstandard>=0.25.0
tqdm>=4.66.0
topicgpt>=0.1.0
openai>=1.0.0
umap-learn>=0.5.0
hdbscan>=0.8.0
